refactor: settings component の命名を PascalCase に統一
section パネル群(account/billing/data/display/profile-settings)と
dialog 群(account-deletion/avatar-change/display-name/email-change/
password-change)を kebab-case から PascalCase に改称し、import元・
stories・testを追随させる。docs内のファイルパス言及も更新。

Refs #1530
