+++ apps/product/src/features/settings/components/AvatarChangeDialog.tsx
