+++ apps/product/src/features/settings/components/PasswordChangeDialog.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
-import { PasswordChangeDialog } from './password-change-dialog';
+import { PasswordChangeDialog } from './PasswordChangeDialog';
 
 const mockUpdateUser = vi.fn();
 const mockSignInWithPassword = vi.fn();
